MIRTH-3549: The Regex and JavaScript attachment handlers now support binary data. When a binary message is received the bytes are simply converted to a Base64 string first and that is used for the incoming message.
MIRTH-2782: The Regex attachment handler now allows multiple regular expressions, each with its own MIME type setting. Migration will happen automatically because the old attachment property names are still supported.

MIRTH-3997: A new "Entire Message" attachment handler was added so that users can easily store the entire inbound message as an attachment. The handler has a single property to set the MIME type.

MIRTH-3998: The MIME type, as well as the inbound/outbound replacement fields, now all support Velocity variable replacement. This applies to the Regex and Entire Message handlers.

MIRTH-3999: Fixed issue where the DICOM attachment viewer would cause the Administrator to hang for invalid DICOM data. We now use a BufferedInputStream and attempt to read at least 128 bytes beforehand to ensure that the Base64 input stream is valid before allowing it to go to the ij plugin.

MIRTH-4001: The Text Viewer now has a "Decode Base64 Data" check box option. It's enabled by default to keep the same behavior, but users may uncheck it to view the raw attachment data (UTF-8 encoded) instead of decoding it as Base64 first.

MIRTH-4002: There is a new dialog that shows up when attachments are viewed in the message browser. Users now have the option of selecting a viewer from the list of available plugins (by default, text/image/DICOM/PDF). A check box exists so that users can opt not to show the dialog and instead have the Administrator always infer the viewer from the MIME type. That is stored as a user preference and may be changed in the Administrator settings. In either case if no valid viewer could be found for a specific MIME type, the dialog will always be shown.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@8086 c801cdd1-710f-0410-a44a-8d9edceeb29b