MIRTH-4006: Fixed issue that caused the same dispatcher ID to be used simultaneously across multiple processing threads. Now instead of just returning -1, a stack (deque actually) is used which contains multiple unique IDs for each possible processing thread. These are kept unique from queue threads because they are negative. In the case that there's only one possible processing thread, or queuing is enabled and there's only one queue thread, 0 is used.
When multiple processing threads get to the dispatcher, Donkey will pop a unique ID off the stack for each thread. LinkedBlockingDeque is a thread-safe implementation so multiple processing threads pushing/popping isn't an issue. The handleSend method will have its own try..finally block, and if a dispatcher ID was popped off then the finally block will push it back on the stack.

This ensures that all processing threads share from a "pool" of unique IDs. We can't use the actual thread IDs in this case because they could come from anywhere and may only be used once.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@8090 c801cdd1-710f-0410-a44a-8d9edceeb29b