MIRTH-3025: Added new Assign To Iterator and Remove From Iterator tasks, and added new functionality to the wizard dialog so that users can choose between creating new or choosing existing Iterators.
The Assign To Iterator task can be used on any step/rule (even Iterators themselves). If one or more Iterators already exist, the dialog will have options to choose between new/existing. When assigning to an iterator, any properties in the step/rule will automatically be modified to inject index variables.

The Remove From Iterator goes the opposite way. It detaches a step/rule from its current parent, and reattaches to the grandparent. Properties in the step/rule will be removed as necessary.

git-svn-id: https://svn.mirthcorp.com/connect/trunk@8200 c801cdd1-710f-0410-a44a-8d9edceeb29b